+++ instrumentation/opentelemetry-instrumentation-requests/tests/test_requests_integration.py
@@ -25,6 +25,13 @@ from opentelemetry import context, trace
 
 # FIXME: fix the importing of this private attribute when the location of the _SUPPRESS_HTTP_INSTRUMENTATION_KEY is defined.
 from opentelemetry.context import _SUPPRESS_HTTP_INSTRUMENTATION_KEY
+from opentelemetry.instrumentation._semconv import (
+    _OTEL_SEMCONV_STABILITY_OPT_IN_KEY,
+    _SPAN_ATTRIBUTES_ERROR_TYPE,
+    _SPAN_ATTRIBUTES_NETWORK_PEER_ADDRESS,
+    _SPAN_ATTRIBUTES_NETWORK_PEER_PORT,
+    _OpenTelemetrySemanticConventionStability,
+)
 from opentelemetry.instrumentation.requests import RequestsInstrumentor
 from opentelemetry.instrumentation.utils import _SUPPRESS_INSTRUMENTATION_KEY
 from opentelemetry.propagate import get_global_textmap, set_global_textmap
@@ -69,12 +76,24 @@ class RequestsIntegrationTestBase(abc.ABC):
     def setUp(self):
         super().setUp()
 
+        test_name = ""
+        if hasattr(self, "_testMethodName"):
+            test_name = self._testMethodName
+        sem_conv_mode = "default"
+        if "new_semconv" in test_name:
+            sem_conv_mode = "http"
+        elif "both_semconv" in test_name:
+            sem_conv_mode = "http/dup"
         self.env_patch = mock.patch.dict(
             "os.environ",
             {
-                "OTEL_PYTHON_REQUESTS_EXCLUDED_URLS": "http://localhost/env_excluded_arg/123,env_excluded_noarg"
+                "OTEL_PYTHON_REQUESTS_EXCLUDED_URLS": "http://localhost/env_excluded_arg/123,env_excluded_noarg",
+                _OTEL_SEMCONV_STABILITY_OPT_IN_KEY: sem_conv_mode,
             },
         )
+
+        _OpenTelemetrySemanticConventionStability._initialized = False
+
         self.env_patch.start()
 
         self.exclude_patch = mock.patch(
@@ -119,6 +138,11 @@ class RequestsIntegrationTestBase(abc.ABC):
         self.assertEqual(span.name, "GET")
 
         self.assertEqual(
+            span.instrumentation_scope.schema_url,
+            "https://opentelemetry.io/schemas/1.11.0",
+        )
+
+        self.assertEqual(
             span.attributes,
             {
                 SpanAttributes.HTTP_METHOD: "GET",
@@ -130,6 +154,84 @@ class RequestsIntegrationTestBase(abc.ABC):
         self.assertIs(span.status.status_code, trace.StatusCode.UNSET)
 
         self.assertEqualSpanInstrumentationInfo(
+            span, opentelemetry.instrumentation.requests
+        )
+
+    def test_basic_new_semconv(self):
+        url_with_port = "http://mock:80/status/200"
+        httpretty.register_uri(
+            httpretty.GET, url_with_port, status=200, body="Hello!"
+        )
+        result = self.perform_request(url_with_port)
+        self.assertEqual(result.text, "Hello!")
+        span = self.assert_span()
+
+        self.assertIs(span.kind, trace.SpanKind.CLIENT)
+        self.assertEqual(span.name, "GET")
+
+        self.assertEqual(
+            span.instrumentation_scope.schema_url,
+            SpanAttributes.SCHEMA_URL,
+        )
+        self.assertEqual(
+            span.attributes,
+            {
+                SpanAttributes.HTTP_REQUEST_METHOD: "GET",
+                SpanAttributes.URL_FULL: url_with_port,
+                SpanAttributes.SERVER_ADDRESS: "mock",
+                _SPAN_ATTRIBUTES_NETWORK_PEER_ADDRESS: "mock",
+                SpanAttributes.HTTP_RESPONSE_STATUS_CODE: 200,
+                SpanAttributes.NETWORK_PROTOCOL_VERSION: "1.1",
+                SpanAttributes.SERVER_PORT: 80,
+                _SPAN_ATTRIBUTES_NETWORK_PEER_PORT: 80,
+            },
+        )
+
+        self.assertIs(span.status.status_code, trace.StatusCode.UNSET)
+
+        self.assertEqualSpanInstrumentationScope(
+            span, opentelemetry.instrumentation.requests
+        )
+
+    def test_basic_both_semconv(self):
+        url_with_port = "http://mock:80/status/200"
+        httpretty.register_uri(
+            httpretty.GET, url_with_port, status=200, body="Hello!"
+        )
+        result = self.perform_request(url_with_port)
+        self.assertEqual(result.text, "Hello!")
+        span = self.assert_span()
+
+        self.assertIs(span.kind, trace.SpanKind.CLIENT)
+        self.assertEqual(span.name, "GET")
+
+        self.assertEqual(
+            span.instrumentation_scope.schema_url,
+            SpanAttributes.SCHEMA_URL,
+        )
+        self.assertEqual(
+            span.attributes,
+            {
+                SpanAttributes.HTTP_METHOD: "GET",
+                SpanAttributes.HTTP_REQUEST_METHOD: "GET",
+                SpanAttributes.HTTP_URL: url_with_port,
+                SpanAttributes.URL_FULL: url_with_port,
+                SpanAttributes.HTTP_HOST: "mock",
+                SpanAttributes.SERVER_ADDRESS: "mock",
+                _SPAN_ATTRIBUTES_NETWORK_PEER_ADDRESS: "mock",
+                SpanAttributes.NET_PEER_PORT: 80,
+                SpanAttributes.HTTP_STATUS_CODE: 200,
+                SpanAttributes.HTTP_RESPONSE_STATUS_CODE: 200,
+                SpanAttributes.HTTP_FLAVOR: "1.1",
+                SpanAttributes.NETWORK_PROTOCOL_VERSION: "1.1",
+                SpanAttributes.SERVER_PORT: 80,
+                _SPAN_ATTRIBUTES_NETWORK_PEER_PORT: 80,
+            },
+        )
+
+        self.assertIs(span.status.status_code, trace.StatusCode.UNSET)
+
+        self.assertEqualSpanInstrumentationScope(
             span, opentelemetry.instrumentation.requests
         )
 
@@ -207,6 +309,57 @@ class RequestsIntegrationTestBase(abc.ABC):
 
         self.assertEqual(
             span.attributes.get(SpanAttributes.HTTP_STATUS_CODE), 404
+        )
+
+        self.assertIs(
+            span.status.status_code,
+            trace.StatusCode.ERROR,
+        )
+
+    def test_not_foundbasic_new_semconv(self):
+        url_404 = "http://mock/status/404"
+        httpretty.register_uri(
+            httpretty.GET,
+            url_404,
+            status=404,
+        )
+        result = self.perform_request(url_404)
+        self.assertEqual(result.status_code, 404)
+
+        span = self.assert_span()
+
+        self.assertEqual(
+            span.attributes.get(SpanAttributes.HTTP_RESPONSE_STATUS_CODE), 404
+        )
+        self.assertEqual(
+            span.attributes.get(_SPAN_ATTRIBUTES_ERROR_TYPE), "404"
+        )
+
+        self.assertIs(
+            span.status.status_code,
+            trace.StatusCode.ERROR,
+        )
+
+    def test_not_foundbasic_both_semconv(self):
+        url_404 = "http://mock/status/404"
+        httpretty.register_uri(
+            httpretty.GET,
+            url_404,
+            status=404,
+        )
+        result = self.perform_request(url_404)
+        self.assertEqual(result.status_code, 404)
+
+        span = self.assert_span()
+
+        self.assertEqual(
+            span.attributes.get(SpanAttributes.HTTP_STATUS_CODE), 404
+        )
+        self.assertEqual(
+            span.attributes.get(SpanAttributes.HTTP_RESPONSE_STATUS_CODE), 404
+        )
+        self.assertEqual(
+            span.attributes.get(_SPAN_ATTRIBUTES_ERROR_TYPE), "404"
         )
 
         self.assertIs(
@@ -368,6 +521,34 @@ class RequestsIntegrationTestBase(abc.ABC):
         )
         self.assertEqual(span.status.status_code, StatusCode.ERROR)
 
+    @mock.patch(
+        "requests.adapters.HTTPAdapter.send",
+        side_effect=requests.RequestException,
+    )
+    def test_requests_exception_new_semconv(self, *_, **__):
+        url_with_port = "http://mock:80/status/200"
+        httpretty.register_uri(
+            httpretty.GET, url_with_port, status=200, body="Hello!"
+        )
+        with self.assertRaises(requests.RequestException):
+            self.perform_request(url_with_port)
+
+        span = self.assert_span()
+        print(span.attributes)
+        self.assertEqual(
+            span.attributes,
+            {
+                SpanAttributes.HTTP_REQUEST_METHOD: "GET",
+                SpanAttributes.URL_FULL: url_with_port,
+                SpanAttributes.SERVER_ADDRESS: "mock",
+                SpanAttributes.SERVER_PORT: 80,
+                _SPAN_ATTRIBUTES_NETWORK_PEER_PORT: 80,
+                _SPAN_ATTRIBUTES_NETWORK_PEER_ADDRESS: "mock",
+                _SPAN_ATTRIBUTES_ERROR_TYPE: "RequestException",
+            },
+        )
+        self.assertEqual(span.status.status_code, StatusCode.ERROR)
+
     mocked_response = requests.Response()
     mocked_response.status_code = 500
     mocked_response.reason = "Internal Server Error"
@@ -456,7 +637,7 @@ class TestRequestsIntegration(RequestsIntegrationTestBase, TestBase):
     @staticmethod
     def perform_request(url: str, session: requests.Session = None):
         if session is None:
-            return requests.get(url)
+            return requests.get(url, timeout=5)
         return session.get(url)
 
     def test_credential_removal(self):
@@ -467,7 +648,7 @@ class TestRequestsIntegration(RequestsIntegrationTestBase, TestBase):
         self.assertEqual(span.attributes[SpanAttributes.HTTP_URL], self.URL)
 
     def test_if_headers_equals_none(self):
-        result = requests.get(self.URL, headers=None)
+        result = requests.get(self.URL, headers=None, timeout=5)
         self.assertEqual(result.text, "Hello!")
         self.assert_span()
 
@@ -489,6 +670,22 @@ class TestRequestsIntergrationMetric(TestBase):
 
     def setUp(self):
         super().setUp()
+        test_name = ""
+        if hasattr(self, "_testMethodName"):
+            test_name = self._testMethodName
+        sem_conv_mode = "default"
+        if "new_semconv" in test_name:
+            sem_conv_mode = "http"
+        elif "both_semconv" in test_name:
+            sem_conv_mode = "http/dup"
+        self.env_patch = mock.patch.dict(
+            "os.environ",
+            {
+                _OTEL_SEMCONV_STABILITY_OPT_IN_KEY: sem_conv_mode,
+            },
+        )
+        self.env_patch.start()
+        _OpenTelemetrySemanticConventionStability._initialized = False
         RequestsInstrumentor().instrument(meter_provider=self.meter_provider)
 
         httpretty.enable()
@@ -496,33 +693,106 @@ class TestRequestsIntergrationMetric(TestBase):
 
     def tearDown(self):
         super().tearDown()
+        self.env_patch.stop()
         RequestsInstrumentor().uninstrument()
         httpretty.disable()
 
     @staticmethod
     def perform_request(url: str) -> requests.Response:
-        return requests.get(url)
+        return requests.get(url, timeout=5)
 
     def test_basic_metric_success(self):
         self.perform_request(self.URL)
 
         expected_attributes = {
-            "http.status_code": 200,
-            "http.host": "examplehost",
-            "net.peer.port": 8000,
-            "net.peer.name": "examplehost",
-            "http.method": "GET",
-            "http.flavor": "1.1",
-            "http.scheme": "http",
+            SpanAttributes.HTTP_STATUS_CODE: 200,
+            SpanAttributes.HTTP_HOST: "examplehost",
+            SpanAttributes.NET_PEER_PORT: 8000,
+            SpanAttributes.NET_PEER_NAME: "examplehost",
+            SpanAttributes.HTTP_METHOD: "GET",
+            SpanAttributes.HTTP_FLAVOR: "1.1",
+            SpanAttributes.HTTP_SCHEME: "http",
         }
 
         for (
             resource_metrics
         ) in self.memory_metrics_reader.get_metrics_data().resource_metrics:
             for scope_metrics in resource_metrics.scope_metrics:
+                self.assertEqual(len(scope_metrics.metrics), 1)
                 for metric in scope_metrics.metrics:
+                    self.assertEqual(metric.unit, "ms")
+                    self.assertEqual(
+                        metric.description,
+                        "measures the duration of the outbound HTTP request",
+                    )
                     for data_point in metric.data.data_points:
                         self.assertDictEqual(
                             expected_attributes, dict(data_point.attributes)
                         )
+                        self.assertEqual(data_point.count, 1)
+
+    def test_basic_metric_new_semconv(self):
+        self.perform_request(self.URL)
+
+        expected_attributes = {
+            SpanAttributes.HTTP_RESPONSE_STATUS_CODE: 200,
+            SpanAttributes.SERVER_ADDRESS: "examplehost",
+            SpanAttributes.SERVER_PORT: 8000,
+            SpanAttributes.HTTP_REQUEST_METHOD: "GET",
+            SpanAttributes.NETWORK_PROTOCOL_VERSION: "1.1",
+        }
+        for (
+            resource_metrics
+        ) in self.memory_metrics_reader.get_metrics_data().resource_metrics:
+            for scope_metrics in resource_metrics.scope_metrics:
+                self.assertEqual(len(scope_metrics.metrics), 1)
+                for metric in scope_metrics.metrics:
+                    self.assertEqual(metric.unit, "s")
+                    self.assertEqual(
+                        metric.description, "Duration of HTTP client requests."
+                    )
+                    for data_point in metric.data.data_points:
+                        self.assertDictEqual(
+                            expected_attributes, dict(data_point.attributes)
+                        )
+                        self.assertEqual(data_point.count, 1)
+
+    def test_basic_metric_both_semconv(self):
+        self.perform_request(self.URL)
+
+        expected_attributes_old = {
+            SpanAttributes.HTTP_STATUS_CODE: 200,
+            SpanAttributes.HTTP_HOST: "examplehost",
+            SpanAttributes.NET_PEER_PORT: 8000,
+            SpanAttributes.NET_PEER_NAME: "examplehost",
+            SpanAttributes.HTTP_METHOD: "GET",
+            SpanAttributes.HTTP_FLAVOR: "1.1",
+            SpanAttributes.HTTP_SCHEME: "http",
+        }
+
+        expected_attributes_new = {
+            SpanAttributes.HTTP_RESPONSE_STATUS_CODE: 200,
+            SpanAttributes.SERVER_ADDRESS: "examplehost",
+            SpanAttributes.SERVER_PORT: 8000,
+            SpanAttributes.HTTP_REQUEST_METHOD: "GET",
+            SpanAttributes.NETWORK_PROTOCOL_VERSION: "1.1",
+        }
+
+        for (
+            resource_metrics
+        ) in self.memory_metrics_reader.get_metrics_data().resource_metrics:
+            for scope_metrics in resource_metrics.scope_metrics:
+                self.assertEqual(len(scope_metrics.metrics), 2)
+                for metric in scope_metrics.metrics:
+                    for data_point in metric.data.data_points:
+                        if metric.unit == "ms":
+                            self.assertDictEqual(
+                                expected_attributes_old,
+                                dict(data_point.attributes),
+                            )
+                        else:
+                            self.assertDictEqual(
+                                expected_attributes_new,
+                                dict(data_point.attributes),
+                            )
                         self.assertEqual(data_point.count, 1)